Move remaining api.types to ApiResult or ApiClient
