--- frontend/lib/api/ApiResponseStatus.ts
@@ -1,5 +1,0 @@
-export enum ApiResponseStatus {
-  Success,
-  ClientError,
-  ServerError,
-}
