+++ frontend/lib/api/ApiResult.ts
@@ -1,7 +1,5 @@
 import { TranslationPath } from "@kiesraad/i18n";
 
-import { ApiResponse, ApiResult } from "./api.types";
-import { ApiResponseStatus } from "./ApiResponseStatus";
 import { ErrorReference } from "./gen/openapi";
 
 export class FatalError extends Error {}
@@ -11,6 +9,12 @@ export class ApiErrorEvent extends Event {
     super("apiError");
     this.error = error;
   }
+}
+
+export enum ApiResponseStatus {
+  Success,
+  ClientError,
+  ServerError,
 }
 
 // Error that should allow the user to retry the request
@@ -62,6 +66,16 @@ export class NotFoundError extends FatalError {
     return this;
   }
 }
+
+export type AnyApiError = ApiError | FatalApiError | NetworkError | NotFoundError;
+
+export interface ApiResponse<T> {
+  status: ApiResponseStatus.Success;
+  code: number;
+  data: T;
+}
+
+export type ApiResult<T, E = AnyApiError> = ApiResponse<T> | E;
 
 export function isError<T>(result: ApiResult<T>): result is Error {
   return (
