+++ frontend/lib/api/index.ts
@@ -2,11 +2,9 @@ export * from "./apportionment";
 export * from "./election";
 export * from "./gen/openapi";
 export * from "./polling_station";
-export type * from "./api.types";
 export * from "./ApiClient";
-export * from "./ApiError";
+export * from "./ApiResult";
 export * from "./ApiProvider";
-export * from "./ApiResponseStatus";
 export * from "./ApiProviderContext";
 export * from "./TestUserProvider";
 export * from "./useApiClient";
